% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{augment.nn_adjust}
\alias{augment.nn_adjust}
\title{Augment data with predicted values}
\usage{
\method{augment}{nn_adjust}(x, new_data, ...)
}
\arguments{
\item{x}{An object of class \code{\link[=nn_adjust]{nn_adjust()}}.}

\item{new_data}{A data frame with the original predictors in their original
format.}

\item{...}{Not currently used.}
}
\value{
The data being predicted with an additional column \code{.pred} that are
the adjusted predictions. If \code{new_data} contains the original outcome column,
there is also a \code{.resid} column.
}
\description{
Augment data with predicted values
}
\examples{
# example code

if (rlang::is_installed(c("ggplot2", "parsnip", "rpart", "MASS"))) {

  library(workflows)
  library(dplyr)
  library(parsnip)
  library(ggplot2)

  # ------------------------------------------------------------------------------
  # Use the 1D motorcycle helmet data as an example

  data(mcycle, package = "MASS")

  # Use every fifth data point as a test point
  in_test <- ( 1:nrow(mcycle) ) \%\% 5 == 0
  cycl_train <- mcycle[-in_test, ]
  cycl_test  <- mcycle[ in_test, ]

  # ------------------------------------------------------------------------------
  # Fit a decision tree

  cart_spec <- decision_tree() \%>\% set_mode("regression")

  cart_fit <-
    workflow(accel ~ times, cart_spec) \%>\%
    fit(data = cycl_train)

  adj_obj <- nn_adjust(cart_fit, cycl_train)\

  # Raw predictions plus data:
  augment(cart_fit, head(cycl_test))

  # Adjusted predictions:
  predict(adj_obj, head(cycl_test), neighbors = 10)

  # Add the data too
  augment(adj_obj, head(cycl_test), neighbors = 10)

}
}
\seealso{
\code{\link[=nn_adjust]{nn_adjust()}}, \code{\link[=predict.nn_adjust]{predict.nn_adjust()}}
}
