% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_adjust.R
\name{nn_adjust}
\alias{nn_adjust}
\alias{nn_adjust.default}
\alias{nn_adjust.workflow}
\title{Nearest Neighbor Adjustment of Predictions}
\usage{
nn_adjust(wflow, training, ...)

\method{nn_adjust}{default}(wflow, training, ...)

\method{nn_adjust}{workflow}(wflow, training, butcher = FALSE, ...)
}
\arguments{
\item{wflow}{A fitted \code{\link[workflows:workflow]{workflows::workflow()}} object.}

\item{training}{A data frame containing the predictors and outcome data used
to create \code{wflow}.}

\item{...}{Not currently used.}

\item{butcher}{A logical: should \code{\link[butcher:butcher]{butcher::butcher()}} be used to trim the
workflow's size?}
}
\value{
An object of class \code{nn_adjust}.
}
\description{
Nearest Neighbor Adjustment of Predictions
}
