% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{augment.nn_adjust}
\alias{augment.nn_adjust}
\title{Augment data with predicted values}
\usage{
\method{augment}{nn_adjust}(x, new_data, ...)
}
\arguments{
\item{x}{An object of class \code{\link[=nn_adjust]{nn_adjust()}}.}

\item{new_data}{A data frame with the original predictors in their original
format.}

\item{...}{Not currently used.}
}
\value{
The data being predicted with an additional column \code{.pred} that are
the adjusted predictions.
}
\description{
Augment data with predicted values
}
\seealso{
\code{\link[=nn_adjust]{nn_adjust()}}, \code{\link[=predict.nn_adjust]{predict.nn_adjust()}}
}
