% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.nn_adjust}
\alias{predict.nn_adjust}
\title{Make adjusted predictions}
\usage{
\method{predict}{nn_adjust}(object, new_data, neighbors = 3, eps = 1/2, cores = 1, ...)
}
\arguments{
\item{object}{An object of class \code{\link[=nn_adjust]{nn_adjust()}}.}

\item{new_data}{A data frame with the original predictors in their original
format.}

\item{neighbors}{An integer for the number of neighbors. Zero indicates no
adjustment.}

\item{eps}{A small constant added to distance to avoid divide by zero.}

\item{cores}{An integer for how many cores \code{\link[gower:gower_topn]{gower::gower_topn()}} should use.}

\item{...}{Not currently used.}
}
\value{
A tibble with a numeric column \code{.pred} that are the adjusted
predictions.
}
\description{
Make adjusted predictions
}
\seealso{
\code{\link[=nn_adjust]{nn_adjust()}}, \code{\link[=augment.nn_adjust]{augment.nn_adjust()}}
}
